% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{class}
\name{coco-class}
\alias{coco-class}
\title{An S4 class to store information}
\description{
An S4 class to store information
}
\section{Slots}{

\describe{
\item{\code{type}}{(\code{character}) One of two available types "dense" or "sparse". See description.}

\item{\code{data}}{(\code{data.frame}) A data.frame with covariates information, where colnames(data) matches model.list specification}

\item{\code{locs}}{(\code{numeric matrix}) a matrix with locs matching data}

\item{\code{z}}{(\code{numeric matrix}) A matrix of dimension n x p with response values}

\item{\code{model.list}}{(\code{list}) A list specifying a model for each aspect of the spatial structure.}

\item{\code{info}}{(\code{list}) a list with information about the coco object}

\item{\code{output}}{(\code{list}) if building an already fitted \code{coco} object (not the standard approach), then requires an output from Optimparallel output, including as well boundaries, etc.}
}}

\author{
Federico Blasi
}
