% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{cocoSim}
\alias{cocoSim}
\title{Marginal and conditional simulation of Gaussian processes with nonstationary covariance function}
\usage{
cocoSim(coco.object, pars, n, seed, standardize, 
type = 'classic', sim.type = NULL, cond.info = NULL)
}
\arguments{
\item{coco.object}{a \code{\link[=coco]{coco()}} object.}

\item{pars}{a vector of parameters values related to \code{model.list}.}

\item{n}{number of realizations to simulate.}

\item{seed}{seed number. default set to NULL.}

\item{standardize}{logical argument describing whether provided covariates
should be standardize (TRUE) or not (FALSE). By default set to TRUE.}

\item{type}{whether parameters are related to a classical parameterization ('classic') or
a difference parameterization \code{'diff'}. Default set to \code{'classic'}. For \code{'sparse'} coco objects, only \code{'diff'} is available.}

\item{sim.type}{if set \code{'cond'} then a conditional simulation takes place.}

\item{cond.info}{a list containing added information to perform conditional simulation.}
}
\value{
a matrix n x dim(data)[1].
}
\description{
simulates a Gaussian process with nonstationary covariance function
from an coco object.
}
\details{
\code{'cond'} sim.type requires specifying in \code{'cond.info'} a list with \code{'newdataset'} a data.frame containing covariates present in model.list at prediction locations,
and \code{'newlocs'} a matrix with locations related to prediction locations, matching indexing of \code{'newdataset'}.
}
\author{
Federico Blasi
}
