% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_rns}
\alias{cov_rns}
\title{Dense covariance function (difference parameterization)}
\usage{
cov_rns(theta, locs, x_covariates, smooth_limits)
}
\arguments{
\item{theta}{vector of parameters}

\item{locs}{a matrix with locations}

\item{x_covariates}{design data.frame}

\item{smooth_limits}{smooth limits}
}
\value{
dense covariance matrix
}
\description{
Dense covariance function (difference parameterization)
}
