% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_rns_classic}
\alias{cov_rns_classic}
\title{Dense covariance function (classic parameterization)}
\usage{
cov_rns_classic(theta, locs, x_covariates)
}
\arguments{
\item{theta}{vector of parameters}

\item{locs}{a matrix with locations}

\item{x_covariates}{design data.frame}
}
\value{
dense covariance matrix with classic parameterization
}
\description{
Dense covariance function (classic parameterization)
}
