% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_rns_pred}
\alias{cov_rns_pred}
\title{Dense covariance function}
\usage{
cov_rns_pred(
  theta,
  locs,
  locs_pred,
  x_covariates,
  x_covariates_pred,
  smooth_limits
)
}
\arguments{
\item{theta}{vector of parameters}

\item{locs}{a matrix with locations}

\item{locs_pred}{a matrix with prediction locations}

\item{x_covariates}{design data.frame}

\item{x_covariates_pred}{design data.frame at prediction locations}

\item{smooth_limits}{smooth limits}
}
\value{
dense covariance matrix
}
\description{
Dense covariance function
}
