% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_rns_taper}
\alias{cov_rns_taper}
\title{Sparse covariance function}
\usage{
cov_rns_taper(
  theta,
  locs,
  x_covariates,
  colindices,
  rowpointers,
  smooth_limits
)
}
\arguments{
\item{theta}{vector of parameters}

\item{locs}{a matrix with locations}

\item{x_covariates}{design data.frame}

\item{colindices}{from spam object}

\item{rowpointers}{from spam object}

\item{smooth_limits}{smooth limits}
}
\value{
sparse covariance matrix between locs and pred_locs
}
\description{
Sparse covariance function
}
