% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cov_rns_taper_pred}
\alias{cov_rns_taper_pred}
\title{Sparse covariance function}
\usage{
cov_rns_taper_pred(
  theta,
  locs,
  locs_pred,
  x_covariates,
  x_covariates_pred,
  colindices,
  rowpointers,
  smooth_limits
)
}
\arguments{
\item{theta}{vector of parameters}

\item{locs}{a matrix with locations}

\item{locs_pred}{a matrix with prediction locations}

\item{x_covariates}{design data.frame}

\item{x_covariates_pred}{design data.frame at prediction locations}

\item{colindices}{from spam object}

\item{rowpointers}{from spam object}

\item{smooth_limits}{smooth limits}
}
\value{
sparse covariance matrix at locs
}
\description{
Sparse covariance function
}
