% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getAIC}
\alias{getAIC}
\title{Retrieve AIC}
\usage{
getAIC(svcov.object)
}
\arguments{
\item{svcov.object}{a fitted svcov object.}
}
\value{
a list with the associated AIC value
}
\description{
Retrieve the Akaike information criterion from a fitted svcov object
}
\author{
Federico Blasi
}
