% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getBIC}
\alias{getBIC}
\title{Retrieve BIC}
\usage{
getBIC(svcov.object)
}
\arguments{
\item{svcov.object}{a fitted svcov object.}
}
\value{
a list with the associated BIC value
}
\description{
Retrieve BIC from a fitted svcov object
}
\author{
Federico Blasi
}
