% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getBoundariesV3}
\alias{getBoundariesV3}
\title{Simple build of boundaries (v3)}
\usage{
getBoundariesV3(coco.object, ...)
}
\arguments{
\item{coco.object}{a coco object}

\item{...}{}
}
\value{
a list with boundaries for the optim L-BFGS-B routine
}
\description{
provides a generic set of upper and lower bounds for the L-BFGS-B routine
}
\author{
Federico Blasi
}
