% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getBoundariesV4}
\alias{getBoundariesV4}
\title{Simple build of boundaries}
\usage{
getBoundariesV4(coco.object, lower.bound, upper.bound)
}
\arguments{
\item{coco.object}{\code{(S4)} a coco.object.}

\item{lower.bound}{\code{(numeric vector)} lower bound for covariate-driven parameters of the nonstationary covariance function.}

\item{upper.bound}{\code{(numeric vector)} upper bound for covariate-driven parameters of the nonstationary covariance function.}
}
\value{
(\code{list}) a list with boundaries and simple init values for the optim L-BFGS-B routine
}
\description{
provides a set of upper and lower bounds for the L-BFGS-B routine
}
\author{
Federico Blasi
}
