% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getCIs}
\alias{getCIs}
\title{Compute Confidence Intervals for an svcov object}
\usage{
getCIs(svcov.object, inv.hess, alpha = 0.05)
}
\arguments{
\item{alpha}{confidence level}

\item{svov.object}{an svcov class (fitted) object}

\item{Hess}{the hessian}
}
\value{
a matrix with lower and upper bound values
}
\description{
Compute confidence intervals for a (fitted) svcov object
}
\author{
Federico Blasi
}
