% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getCondNumber}
\alias{getCondNumber}
\title{Computes the condition number of the associated correlation matrix of the fitted svcov object}
\usage{
getCondNumber(svcov.object)
}
\arguments{
\item{svov.object}{an svcov class (fitted) object}
}
\value{
the condition number
}
\description{
Compute the trend of the (fitted) svcov object
}
\author{
Federico Blasi
}
