% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getCRPS}
\alias{getCRPS}
\title{Based on a set of predictions retrieves the Logrank}
\usage{
getCRPS(z.pred, mean.pred, sd.pred)
}
\arguments{
\item{z.pred}{\code{(numeric vector)}.}

\item{mean.pred}{\code{(numeric vector)}.}

\item{sd.pred}{\code{(numeric vector)}.}
}
\value{
retrieves CRPS
}
\description{
Retrieves the estimated spatial effects of the spatial structure
}
\author{
Federico Blasi
}
