% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getDensityFromDelta}
\alias{getDensityFromDelta}
\title{Based on a specific taper scale (delta), retrieves the density of the covariance matrix.}
\usage{
getDensityFromDelta(coco.object, delta)
}
\arguments{
\item{coco.object}{\code{(S4)} a fitted \code{\link[=coco]{coco()}} object.}

\item{delta}{\code{(numeric)} a delta taper scale (delta).}
}
\value{
(\code{numeric vector}) the associate density of the tapered covariance matrix.
}
\description{
Based on a specific taper scale (delta), retrieves the density of the covariance matrix.
}
\author{
Federico Blasi
}
