% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getDesignMatrix}
\alias{getDesignMatrix}
\title{Create an efficient design matrix based on a list of aspect models}
\usage{
getDesignMatrix(model.list, data)
}
\arguments{
\item{model.list}{\code{(list)} a list of formulas, one for each source of nonstationarity, specifying the
models.}

\item{data}{\code{(data.frame)} a data.frame.}
}
\value{
(\code{list}) a list with two elements: a design matrix of dimension
(n x p), and a par.pos object, indexing columns of the design matrix to each of the spatially-varying functions.
}
\description{
Creates a unique design matrix based on model specification for
each of the different potentially spatially varying aspects.
}
\author{
Federico Blasi
}
