% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getEstims}
\alias{getEstims}
\title{Retrieve estimates from a fitted svcov object}
\usage{
getEstims(svcov.object)
}
\arguments{
\item{svcov.object}{a fitted svcov object.}
}
\value{
a list with the estimates parameters for the different aspects
}
\description{
Retrieve estimates from a fitted svcov object
}
\author{
Federico Blasi
}
