% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getHessian}
\alias{getHessian}
\title{getHessian}
\usage{
getHessian(coco.object, ncores = parallel::detectCores() - 1, 
eps = .Machine$double.eps^(1/4))
}
\arguments{
\item{coco.object}{a fitted coco object}

\item{ncores}{number of cores used for the computation}

\item{eps}{...}
}
\value{
a symmetric matrix pxp of the approximated (observed) Hessian
}
\description{
returns the approximate (observed) Hesian (inverse of Fisher Information Matrix)
}
\author{
Federico Blasi
}
