% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getLoglik}
\alias{getLoglik}
\title{Retrieve the loglikelihood value}
\usage{
getLoglik(svcov.object)
}
\arguments{
\item{svov.object}{an svcov class (fitted) object}
}
\value{
loglikelihood value of the OptimParallel function
}
\description{
Retrieve the loglikelihood value from a fitted svcov object.
}
\author{
Federico Blasi
}
