% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getLogRank}
\alias{getLogRank}
\title{Based on a set of predictions retrieves the Logrank}
\usage{
getLogRank(z.pred, mean.pred, sd.pred)
}
\arguments{
\item{z.pred}{...}

\item{mean.pred}{...}

\item{sd.pred}{...}
}
\value{
retrieves LogRank
}
\description{
Retrieves the estimated spatial effects of the spatial structure
}
\author{
Federico Blasi
}
