% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getLogScore}
\alias{getLogScore}
\title{Based on a set of predictions retrieves the LogScore}
\usage{
getLogScore(z.pred, mean.pred, sd.pred)
}
\arguments{
\item{z.pred}{\code{(numeric vector)}.}

\item{mean.pred}{\code{(numeric vector)}.}

\item{sd.pred}{\code{(numeric vector)}.}
}
\value{
retrieves LogScore
}
\description{
Retrieves the estimated spatial effects of the spatial structure
}
\author{
Federico Blasi
}
