% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getModHess}
\alias{getModHess}
\title{Retrieves the modified inverse of the hessian}
\usage{
getModHess(svcov.object, inv.hess, alpha = 0.05)
}
\arguments{
\item{inv.hess}{Inverse of the Hessian}

\item{svov.object}{an svcov class (fitted) object}

\item{alpha}{confidence level}
}
\value{
a matrix with lower and upper bound values
}
\description{
Compute confidence intervals for a (fitted) svcov object
}
\author{
Federico Blasi
}
