% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg2loglikelihood.R
\name{GetNeg2loglikelihood}
\alias{GetNeg2loglikelihood}
\title{GetNeg2loglikelihood}
\usage{
GetNeg2loglikelihood(theta, par.pos, locs, x_covariates, 
smooth.limits, z, n, lambda)
}
\arguments{
\item{theta}{\code{(numeric vector)} a vector with parameters values.}

\item{par.pos}{\code{(list)} par.pos list.}

\item{locs}{\code{(matrix)} spatial location matrix.}

\item{x_covariates}{\code{(data.frame)} design matrix.}

\item{smooth.limits}{\code{(numeric vector)} smooth.limits.}

\item{z}{\code{(numeric vector)} a vector of observed values.}

\item{n}{\code{(integer)} dim(z)[1].}

\item{lambda}{\code{(numeric)} regularization parameter.}
}
\value{
value
}
\description{
compute the negative 2 log likelihood based on theta
}
\author{
Federico Blasi
}
