% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg2loglikelihood.R
\name{GetNeg2loglikelihoodProfile}
\alias{GetNeg2loglikelihoodProfile}
\title{GetNeg2loglikelihoodProfile}
\usage{
function(theta)
}
\arguments{
\item{theta}{a vector with parameters values}

\item{par.pos}{a vector with parameters values}

\item{locs}{a vector with parameters values}

\item{x_covariates}{a vector with parameters values}

\item{smooth.limits}{a vector with parameters values}

\item{n}{a vector with parameters values}

\item{z}{a vector with parameters values}

\item{lambda}{amount of regularization}
}
\value{
neg2loglikelihood
}
\description{
compute the negative 2 log likelihood based on theta
}
\examples{
rnorm(1)
}
\author{
Federico Blasi
}
