% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg2loglikelihood.R
\name{GetNeg2loglikelihoodREML}
\alias{GetNeg2loglikelihoodREML}
\title{GetNeg2loglikelihoodREML}
\usage{
GetNeg2loglikelihoodREML(theta, par.pos, locs, x_covariates, x_betas,
smooth.limits, z, n, lambda, safe = TRUE)
}
\arguments{
\item{theta}{\code{(numeric vector)} a vector with parameters values.}

\item{par.pos}{\code{(list)} par.pos list.}

\item{locs}{\code{(matrix)} spatial location matrix.}

\item{x_covariates}{\code{(data.frame)} design matrix.}

\item{x_betas}{\code{(matrix) or (data.frame)} design matrix for the spatial mean.}

\item{smooth.limits}{\code{(numeric vector)} smooth.limits.}

\item{z}{\code{(numeric vector)} a vector of contrasts.}

\item{n}{\code{(integer)} dim(z)[1].}

\item{lambda}{\code{(numeric vector)} a vector including lambda.Sigma, lambda.betas, lambda.reg (in this order)}

\item{safe}{\code{(TRUE/FALSE)} if \code{TRUE} returns a large pre-defined value under Cholesky errors. Default \code{TRUE}.}
}
\value{
value
}
\description{
compute the negative 2 log REML likelihood based on theta
}
\author{
Federico Blasi
}
