% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getPen}
\alias{getPen}
\title{Returns the penalization term}
\usage{
getPen(n, lambda, theta_list, smooth.limits)
}
\arguments{
\item{n}{\code{(integer)}.}

\item{lambda}{\code{(numeric)}.}

\item{theta_list}{\code{(list)}.}

\item{smooth.limits}{\code{(numeric vector)}.}
}
\value{
(\code{numeric}) retrieves penalization term.
}
\description{
Returns the penalization term.
}
\author{
Federico Blasi
}
