% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getSpatMean}
\alias{getSpatMean}
\title{Computes the spatial mean of a (fitted) coco object}
\usage{
getSpatMean(coco.object)
}
\arguments{
\item{coco.object}{\code{(S4)} a fitted coco S4 object.}
}
\value{
(\code{numeric vector}) a vector with the adjusted trend.
}
\description{
Computes the spatial mean of the (fitted) coco object.
}
\author{
Federico Blasi
}
