% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getTrend}
\alias{getTrend}
\title{Computes the spatial trend of a (fitted) coco object}
\usage{
getTrend(coco.object)
}
\arguments{
\item{coco.object}{a fitted \code{\link[=coco]{coco()}} object.}
}
\value{
a vector with the adjusted trend.
}
\description{
Compute the trend of the (fitted) coco object
}
\author{
Federico Blasi
}
