% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holes}
\alias{holes}
\title{Holes Data Set}
\format{
A data frame with rows and variables:
\describe{
\item{var1}{Description of var1}
\item{var2}{Description of var2}
}
}
\source{
Source of the data
}
\usage{
holes
}
\description{
Description of the holes data set.
}
\examples{
data(holes)
}
\keyword{datasets}
