% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFunctions.R
\name{is.formula}
\alias{is.formula}
\title{check whether an R object is a formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{(ANY) an R object.}
}
\value{
TRUE/FALSE
}
\description{
check whether an R object is a formula
}
\author{
Federico Blasi
}
