% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot,coco,missing-method}
\alias{plot,coco,missing-method}
\alias{plot,coco-method}
\title{Plot Method for coco objects}
\usage{
\S4method{plot}{coco,missing}(x, y, type = NULL, index = NULL, factr = 0.1, ...)
}
\arguments{
\item{x}{(\code{S4}) A fitted object of class \code{coco}.}

\item{y}{Not used.}

\item{type}{(\code{character}  or \code{NULL}) The type of plot. NULL or "ellipse" for drawing ellipse of the convolution kernels.}

\item{index}{(\code{integer vector}) For plotting local correlation plots.}

\item{factr}{(\code{numeric}) Factor rate for size of ellipses.}

\item{...}{Additional arguments passed to \link[fields]{quilt.plot}.}
}
\value{
Several plots are created.
}
\description{
This method plots objects of class \code{coco}.
}
\author{
Federico Blasi
}
