% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOptimInfo}
\alias{plotOptimInfo}
\title{Plot log info detailed}
\usage{
plotOptimInfo(svcov.object, ...)
}
\arguments{
\item{svcov.object}{an optimized svcov.object}

\item{...}{arguments for par()}
}
\value{
Outputs a sequence of plots detailing parameters during the 
optimization routine
}
\description{
plot output of optim
}
\author{
Federico Blasi
}
