% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVar}
\alias{plotVar}
\title{Plot variable}
\usage{
plotOptimInfo(svcov.object, ...)
}
\arguments{
\item{...}{arguments for par()}

\item{svcov.object}{an optimized svcov.object}
}
\value{
Outputs a sequence of plots detailing parameters during the 
optimization routine
}
\description{
plot output of optim
}
\author{
Federico Blasi
}
