% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,coco-method}
\title{Print Method for Coco Class}
\usage{
\S4method{print}{coco}(x, inv.hess = NULL, ...)
}
\arguments{
\item{x}{An object of class 'coco'.}

\item{inv.hess}{inverse of the approximated hessian matrix (getHessian)}

\item{...}{Additional arguments to be passed to plot.}
}
\value{
print the coco object
}
\description{
This method prints objects of class 'coco'.
}
\author{
Federico Blasi
}
