% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,coco-method}
\title{Show Method for Coco Class}
\usage{
\S4method{show}{coco}(object)
}
\arguments{
\item{object}{An object of class 'coco'.}
}
\value{
A plot is created.
}
\description{
This method show objects of class 'coco'.
}
\author{
Federico Blasi
}
