% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stripes}
\alias{stripes}
\title{Stripes Data Set}
\format{
A list with training and test data.frame with rows and variables:
\describe{
\item{x}{first spatial coordinate}
\item{y}{second spatial coordinate}
\item{cox_x}{first spatial characteristic}
\item{cov_y}{second spatial characteristic}
\item{cov_xy}{third spatial characteristic}
\item{z}{response variable}
}
}
\source{
Source of the data
}
\usage{
stripes
}
\description{
The synthetic "stripes" provides a set of training and test data.frame of a Gaussian process realization with a (inherently sparse) nonstationary covariance function.
Several stripes are present in the training dataset, and the task is to predict them.
}
\examples{
data(stripes)
}
\keyword{datasets}
