% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,coco-method}
\title{Summary Method for Coco Class}
\usage{
\S4method{summary}{coco}(object, inv.hess = NULL, ...)
}
\arguments{
\item{object}{An object of class 'coco'.}

\item{inv.hess}{inverse of the approximated hessian matrix (getHessian)}

\item{...}{Additional arguments to be passed to plot.}
}
\value{
summary the coco object
}
\description{
This method prints objects of class 'coco'.
}
\author{
Federico Blasi
}
