% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumsmoothlone}
\alias{sumsmoothlone}
\title{smoothed-L1 penalization over the covariate-driven parameters}
\usage{
sumsmoothlone(x, lambda, alpha = 1e+06)
}
\arguments{
\item{x}{a vector of parameters to penalize}

\item{lambda}{penalization hyperparameter}

\item{alpha}{hyperparameter to approximate L1 penalization. By default 1e6 provides a decent approximation to L1, while also being a differentiable function}
}
\value{
penalization
}
\description{
smoothed-L1 penalization over the covariate-driven parameters
}
