% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcov.R
\name{svcov}
\alias{svcov}
\title{svcov Class}
\usage{
svcov(type, model.list, data, locs, z, 
taper = function(){}, delta = NA_real_, output = list())
}
\value{
Creates an svcov object
}
\description{
Creates a svcov object
}
\section{Slots}{

\describe{
\item{\code{type}}{One of two available types "dense" or "sparse". See description.}

\item{\code{model.list}}{A list specyfing a model for each aspect of the spatial structure.}

\item{\code{data}}{A data.frame with covariates information, where colnames(data) matches model.list specification}

\item{\code{locs}}{a matrix with locs matching data}

\item{\code{z}}{A vector with response values}

\item{\code{taper}}{If "sparse", a compact supported taper function from spam package}

\item{\code{delta}}{a taper range}

\item{\code{output}}{empty or an output from optimparallel output, including as well boundaries}
}}

\seealso{
[spam::cov.wend1()]
}
\author{
Federico Blasi
}
