% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{svcovOptim}
\alias{svcovOptim}
\title{Optimizer for object class svcov}
\usage{
svcovOptim(svcov.object, boundaries = list(), ncores = parallel::detectCores())
}
\arguments{
\item{svcov.object}{a svcov object. See ?svcov()}

\item{boundaries}{if provided, a list with lower, init, and upper values. 
if not is computed based on generic fast_init_boundaries()}

\item{ncores}{number of cores for the optimization routine.}

\item{optim.control}{number of cores for the optimization routine.}

\item{optim.type}{Optimization approach.}
}
\value{
a svcov object with an updated output slot, with extra information 
with boundaries information
}
\description{
Optimizer based on Optimparallel L-BFGS-B optimizier for svcov class.
}
\author{
Federico Blasi
}
