% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{svcovPredict}
\alias{svcovPredict}
\title{Prediction for object class svcov}
\usage{
svcovPredict(x, newdataset = newdataset, newlocs = newlocs, type = 'mean')
}
\arguments{
\item{svcov.object}{a svcov object}

\item{newdataset}{a data.frame including covariates present in model.list}

\item{newlocs}{a matrix with locations related to newdataset}

\item{type}{whether "mean" or "pred", which gives a point prediction as well as the prediction uncertainty}
}
\value{
a list with mean predictions and CRPS values, latter if specified.
}
\description{
Prediction for a fitted svcov object.
}
\author{
Federico Blasi
}
