% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{svcovSim}
\alias{svcovSim}
\title{Marginal simulation of Gaussian Processes with nonstationary covariance function}
\usage{
svcovSim(svcov.object, pars, n, seed, standardize, type = 'classic')
}
\arguments{
\item{svcov.object}{a vector of length p, where p is the number of parameters for}

\item{pars}{a vector of length p, where p is the number of parameters for
each of the models}

\item{n}{number of realizations to simulate}

\item{seed}{seed number. defalt set to NULL.}

\item{standardize}{logical argument describing whether provided covariates
should be standardize (TRUE) or not (FALSE). By default set to TRUE}

\item{type}{wether parameters are related to a classical parameterization ('classic') or
a difference parameterization 'diff' . Default set to 'classic'.}

\item{sim.type}{if set 'cond' then a conditional simulation takes place.}

\item{cond.info}{a list containing information to perform a conditional simulation.}
}
\value{
a list() with the structure needed to fit cov.rns functions
}
\description{
simulates a Gaussian Process with nonstationary covariance function
from an svcov object.
}
\examples{
model.list <- list("std.dev" = as.formula(" ~ 1 + lati_s * long_s"),
                      "scale" = as.formula(" ~ 1 + elev_s"),
                      "aniso" = as.formula(" ~ 1 + elev_s"),
                      "tilt" = as.formula(" ~ 1 + elev_s"),
                      "smooth" = as.formula(" ~ 1"),
                      "nugget" = -Inf)
}
\author{
Federico Blasi
}
