% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{class}
\name{coco-class}
\alias{coco-class}
\title{An S4 class to store information}
\description{
An S4 class to store information
}
\section{Slots}{

\describe{
\item{\code{type}}{One of two available types "dense" or "sparse". See description.}

\item{\code{data}}{A data.frame with covariates information, where colnames(data) matches model.list specification}

\item{\code{locs}}{a matrix with locs matching data}

\item{\code{z}}{A vector with response values}

\item{\code{model.list}}{A list specyfing a model for each aspect of the spatial structure.}

\item{\code{info}}{a list with information about the coco object}

\item{\code{output}}{an output from optimparallel output, including as well boundaries
information as another element of the list}
}}

\author{
Federico Blasi
}
