% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocons.R
\name{coco}
\alias{coco}
\title{coco Class}
\usage{
coco(type, data, locs, z, model.list, info, output = list())
}
\arguments{
\item{type}{One of two available types \code{'dense'} or \code{'sparse'}. See description.}

\item{data}{A data.frame with covariates information, where colnames(data) matches model.list specification.}

\item{locs}{A matrix with locs matching data.}

\item{z}{A matrix of n x r response realizations, one realization per column. When considering only one realization, a vector can also be provided.}

\item{model.list}{A list specifying a model for each aspect of the spatial structure.}

\item{info}{A list specifying characteristics of the coco object.}

\item{output}{Empty or an output from optimParallel output, including as well boundaries.}
}
\value{
Creates a coco object.
}
\description{
Creates a coco S4 object.
}
\details{
This S4 class is the centerpiece of the \pkg{cocons} package. Two types of coco objects can be created. One is a \code{'dense'} type, meaning that the associated nonstationary covariance functions are dense, or
a \code{'sparse'} object, which, in combination with the Tapering approach, induce zeroes in the covariance matrix to make it sparse and to unlock a set of efficient algorithms to speed up estimation and prediction routines.

An important component of the coco S4 class is the \code{model.list} specification, involving different formulas provided as a list, where each of them specifies a covariate-based parametric model for a specific source of nonstationarity. It involves \code{"trend"} for the spatial trend,
the \code{"std. dev"} for the marginal standard deviation, \code{"scale"}, \code{"aniso"} and \code{"tilt"}, each of them shaping specific aspects of the local spatial geometrically anisotropy structure,
\code{"smooth"} handling local smoothness, and \code{"nugget"} handling the local nugget effect.

Lastly, arguments for the \code{'info'} list argument involve:
\itemize{
\item \code{'lambda'}: a positive scalar specifying the regularization parameter.
\item \code{'smooth.limits'}: specifying the allowed range of variation for the spatially varying smoothness.
\item \code{'taper'}: specifying the desired taper function from the spam package (only for 'sparse' coco objects).
\item \code{'delta'}: specifying the taper range/scale (only for 'sparse' coco objects).
\item \code{'cat.vars'}: index of those variables in \code{data} that should not be scaled during the optimization (e.g., categorical).
}
}
\seealso{
\code{\link[spam:covmat]{spam::cov.wend1()}}
}
\author{
Federico Blasi
}
