% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocons.R
\name{coco}
\alias{coco}
\title{coco Class}
\usage{
coco(type, data, locs, z, model.list, info, output = list())
}
\arguments{
\item{type}{One of two available types "dense" or "sparse". See description.}

\item{data}{A data.frame with covariates information, where colnames(data) matches model.list specification}

\item{locs}{a matrix with locs matching data}

\item{z}{A vector with response values}

\item{model.list}{A list specyfing a model for each aspect of the spatial structure.}

\item{info}{A list specifying characteristics of the coco object}

\item{output}{empty or an output from optimparallel output, including as well boundaries}
}
\value{
Creates a coco object
}
\description{
Creates a coco object
}
\seealso{
\code{\link[spam:covmat]{spam::cov.wend1()}}
}
\author{
Federico Blasi
}
