\name{cocons-package}
\alias{cocons-package}
\alias{cocons}
\docType{package}
\title{Covariate-based Covariance Functions for Nonstationary Gaussian Processes}
\description{
Estimation and prediction of nonstationary Gaussian processes with modular covariate-based covariance functions. Routines for handling large datasets are also provided.
}
\author{
Federico Blasi [aut, cre], \email{federico.blasi@math.uzh.ch},
Reinhard Furrer [ctb]
}