% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{cocoOptim}
\alias{cocoOptim}
\title{Optimizer for object class coco}
\usage{
cocoOptim(coco.object, boundaries = list(), 
ncores = parallel::detectCores(), optim.control, optim.type,...)
}
\arguments{
\item{coco.object}{a coco object. See ?coco()}

\item{boundaries}{if provided, a list with lower, init, and upper values.
if not is computed based on generic fast_init_boundaries()}

\item{ncores}{number of cores for the optimization routine.}

\item{optim.control}{number of cores for the optimization routine.}

\item{optim.type}{Optimization approach.}

\item{...}{extra arguments passed to the optimparallel function}
}
\value{
a coco object with an updated output slot, with extra information
with boundaries information
}
\description{
Optimizer based on Optimparallel L-BFGS-B optimizier for coco class.
}
\author{
Federico Blasi
}
