% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{cocoPredict}
\alias{cocoPredict}
\title{Prediction for object class coco}
\usage{
cocoPredict(coco.object, newdataset, newlocs, type = 'mean')
}
\arguments{
\item{coco.object}{a coco object}

\item{newdataset}{a data.frame including covariates present in model.list}

\item{newlocs}{a matrix with locations related to newdataset}

\item{type}{whether "mean" or "pred", which gives a point prediction as well as the prediction uncertainty}
}
\value{
a list with mean predictions and CRPS values, latter if specified.
}
\description{
Prediction for a fitted coco object.
}
\author{
Federico Blasi
}
