% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{cocoPredict}
\alias{cocoPredict}
\title{Prediction for object class coco}
\usage{
cocoPredict(coco.object, newdataset, newlocs, type = 'mean', ...)
}
\arguments{
\item{coco.object}{a fitted coco object}

\item{newdataset}{a data.frame including covariates present in model.list at prediction locations}

\item{newlocs}{a matrix with locations related to prediction locations, matching indexing of newdataset}

\item{type}{whether "mean" or "pred", which gives a point prediction for the former, as well as a combination of point prediction as well as prediction uncertainty for the latter}

\item{...}{when coco.object has multiple realizations, specifying 'index.pred' specifying which column of coco.object\@z should be used to perform predictions}
}
\value{
a list with trend, and mean predictions and uncertainty quantification, if 'pred' is specified.
}
\description{
Prediction for a fitted coco object.
}
\author{
Federico Blasi
}
