% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getAIC}
\alias{getAIC}
\title{Retrieve AIC}
\usage{
getAIC(coco.object)
}
\arguments{
\item{coco.object}{\code{(S4)} a fitted coco S4 object.}
}
\value{
(\code{numeric}) the associated AIC value
}
\description{
Retrieve the Akaike information criterion from a fitted coco object.
}
\author{
Federico Blasi
}
