% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getBIC}
\alias{getBIC}
\title{Retrieve BIC}
\usage{
getBIC(coco.object)
}
\arguments{
\item{coco.object}{\code{(S4)} a fitted coco S4 object.}
}
\value{
(\code{numeric}) the associated BIC value
}
\description{
Retrieve BIC from a fitted coco object.
}
\author{
Federico Blasi
}
