% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getBoundaries}
\alias{getBoundaries}
\title{Simple build of boundaries}
\usage{
getBoundaries(x, lower.value, upper.value)
}
\arguments{
\item{x}{\code{(S4) or (list)} a coco.object or a par.pos list (as output from \link{getDesignMatrix})}

\item{lower.value}{\code{(numeric vector)} if provided, provides a vector filled with values lower.value.}

\item{upper.value}{\code{(numeric vector)} if provided, provides a vector filled with values upper.value.}
}
\value{
(\code{list}) a list with boundaries and simple init values for the optim L-BFGS-B routine
}
\description{
provides a generic set of upper and lower bounds for the L-BFGS-B routine
}
\author{
Federico Blasi
}
