% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getBoundariesV2}
\alias{getBoundariesV2}
\title{Simple build of boundaries (v2)}
\usage{
getBoundariesV2(coco.object, lower.value, upper.value)
}
\arguments{
\item{coco.object}{a coco object}

\item{param.limits}{a vector of c(lower,init,upper) values for the associated param.}
}
\value{
a list with boundaries for the optim L-BFGS-B routine
}
\description{
provides a generic set of upper and lower bounds for the L-BFGS-B routine
}
\author{
Federico Blasi
}
