% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getBoundariesV2}
\alias{getBoundariesV2}
\title{Simple build of boundaries (v2)}
\usage{
getBoundariesV2(coco.object, mean.limits, std.dev.limits, 
scale.limits, aniso.limits, tilt.limits, smooth.limits, nugget.limits)
}
\arguments{
\item{coco.object}{\code{(S4)} a coco object.}

\item{mean.limits}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{std.dev.limits}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{scale.limits}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{aniso.limits}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{tilt.limits}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{smooth.limits}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{nugget.limits}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}
}
\value{
(\code{list}) a list with boundaries for the optim L-BFGS-B routine
}
\description{
provides a generic set of upper and lower bounds for the L-BFGS-B routine
}
\author{
Federico Blasi
}
