% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getBoundariesV3}
\alias{getBoundariesV3}
\title{Simple build of boundaries (v3)}
\usage{
getBoundariesV3(coco.object, mean.limits, global.lower, 
std.dev.max.effects, 
scale.max.effects, aniso.max.effects, tilt.max.effects, 
smooth.max.effects, nugget.max.effects)
}
\arguments{
\item{coco.object}{\code{(S4)} a coco object.}

\item{mean.limits}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{global.lower}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{std.dev.max.effects}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{scale.max.effects}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{aniso.max.effects}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{tilt.max.effects}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{smooth.max.effects}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}

\item{nugget.max.effects}{\code{(numeric vector)} a vector of c(lower,init,upper) values for the associated param.}
}
\value{
(\code{list}) a list with boundaries for the optim L-BFGS-B routine
}
\description{
provides a generic set of upper and lower bounds for the L-BFGS-B routine
}
\author{
Federico Blasi
}
