% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getCIs}
\alias{getCIs}
\title{Compute Confidence Intervals for an coco object}
\usage{
getCIs(coco.object, inv.hess, alpha = 0.05)
}
\arguments{
\item{coco.object}{a coco class (fitted) object}

\item{inv.hess}{Inverse of the Hessian}

\item{alpha}{confidence level}
}
\value{
a matrix with confidence intervals for each parameter in the model
}
\description{
Compute confidence intervals for a (fitted) coco object
}
\author{
Federico Blasi
}
