% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getCondNumber}
\alias{getCondNumber}
\title{Condition number for (fitted) coco objects}
\usage{
getCondNumber(coco.object)
}
\arguments{
\item{coco.object}{\code{(S4)} a fitted coco S4 object.}
}
\value{
(\code{numeric}) the condition number.
}
\description{
Computes the condition number of the associated correlation matrix of the fitted coco object.
}
\author{
Federico Blasi
}
