% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getCondNumber}
\alias{getCondNumber}
\title{Computes the condition number of the associated correlation matrix of the fitted coco object}
\usage{
getCondNumber(coco.object)
}
\arguments{
\item{coco.object}{a coco class (fitted) object}
}
\value{
the condition number
}
\description{
Compute the trend of the (fitted) coco object
}
\author{
Federico Blasi
}
