% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getCovMatrix}
\alias{getCovMatrix}
\title{Covariance matrix from a fitted coco object}
\usage{
getCovMatrix(coco.object, type = 'global', index = NULL)
}
\arguments{
\item{coco.object}{a fitted \code{\link[=coco]{coco()}} object.}

\item{type}{whether \code{'global'} to retrieve the regular covariance matrix, or \code{'local'} to retrieve global covariance
based on the local aspects of a specific location (not implemented yet)}

\item{index}{index to perform local covariance matrix (not implemented yet)}
}
\value{
a covariance matrix
}
\description{
Retrieves the associated covariance matrix from a fitted coco object
}
\author{
Federico Blasi
}
