% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getCovMatrix}
\alias{getCovMatrix}
\title{Covariance matrix for "coco" class}
\usage{
getCovMatrix(coco.object, type = 'global', index = NULL)
}
\arguments{
\item{coco.object}{\code{(S4)} a fitted \code{\link[=coco]{coco()}} object.}

\item{type}{\code{(character)} whether \code{'global'} to retrieve the regular covariance matrix, or \code{'local'} to retrieve global covariance.
based on the local aspects of a specific location (not implemented yet).}

\item{index}{\code{(integer)} index to perform local covariance matrix (not implemented yet).}
}
\value{
(\code{matrix}) a n x n covariance matrix.
}
\description{
Compute the covariance matrix of \code{coco.object}.
}
\author{
Federico Blasi
}
