% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getCRPS}
\alias{getCRPS}
\title{Based on a set of predictions retrieves the Logrank}
\usage{
getCRPS(svcov.object, type = 'global', index = NULL)
}
\arguments{
\item{svov.object}{an svcov class (fitted) object}

\item{type}{whether 'global' to retrieve the regular covariance matrix, or 'local' to retrieve global covariance 
based on the local aspect of a specific location (not implemented yet)}

\item{index}{index to perform local covariance matrix (not implemented yet)}
}
\value{
a vector with the adjusted trend
}
\description{
Retrieves the estimated spatial effects of the spatial structure
}
\author{
Federico Blasi
}
