% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getCRPS}
\alias{getCRPS}
\title{Based on a set of predictions retrieves the Logrank}
\usage{
getCRPS(z.pred, mean.pred, sd.pred)
}
\arguments{
\item{z.pred}{\code{(numeric vector)}.}

\item{mean.pred}{\code{(numeric vector)}.}

\item{sd.pred}{\code{(numeric vector)}.}
}
\value{
(\code{numeric vector}) retrieves CRPS.
}
\description{
Retrieves the Continuous Ranked Probability Score (CRPS) [1].
}
\references{
[1] Gneiting, Tilmann, and Adrian E. Raftery. \emph{"Strictly proper scoring rules, prediction, and estimation."} Journal of the American statistical Association 102.477 (2007): 359-378.
}
\author{
Federico Blasi
}
