% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getEstims}
\alias{getEstims}
\title{Retrieve estimates from a fitted coco object}
\usage{
getEstims(coco.object)
}
\arguments{
\item{coco.object}{\code{(S4)} a fitted coco S4 object.}
}
\value{
(\code{list}) a list with the estimates parameters for the different aspects
}
\description{
Retrieve estimates from a fitted coco object.
}
\author{
Federico Blasi
}
