% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getLoglik}
\alias{getLoglik}
\title{Retrieve the loglikelihood value}
\usage{
getLoglik(coco.object)
}
\arguments{
\item{coco.object}{\code{(S4)} a fitted coco S4 object.}
}
\value{
(\code{numeric}) wrap for value from a OptimParallel object
}
\description{
Retrieve the loglikelihood value from a fitted coco object.
}
\author{
Federico Blasi
}
