% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getLogScore}
\alias{getLogScore}
\title{Computes the Log-Score}
\usage{
getLogScore(z.pred, mean.pred, sd.pred)
}
\arguments{
\item{z.pred}{\code{(numeric vector)}.}

\item{mean.pred}{\code{(numeric vector)}.}

\item{sd.pred}{\code{(numeric vector)}.}
}
\value{
(\code{numeric vector}) retrieves Log-Score.
}
\description{
Retrieves the Log-Score [1].
}
\references{
[1] Gneiting, Tilmann, and Adrian E. Raftery. \emph{"Strictly proper scoring rules, prediction, and estimation."} Journal of the American statistical Association 102.477 (2007): 359-378.
}
\author{
Federico Blasi
}
