% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getModelLists}
\alias{getModelLists}
\title{Builds the necessary input for building covariance matrices}
\usage{
getModelLists(theta, par.pos, type = 'diff')
}
\arguments{
\item{theta}{a vector of length p, where p is the number of parameters for
each of the models}

\item{par.pos}{a list detailing in which position of each aspect the elements
of theta should be placed. Expected to be output of getDesignMatrix}

\item{type}{whether parameters are related to a classical parameterization ('classic') or
a difference parameterization 'diff' . Default set to 'diff'.}
}
\value{
a list() of different spatial aspects and mean required for the cov.rns functions
}
\description{
Returns a list of parameter vectors for each of the aspects.
}
\author{
Federico Blasi
}
