% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getModHess}
\alias{getModHess}
\title{Retrieves the modified inverse of the hessian}
\usage{
getModHess(coco.object, inv.hess)
}
\arguments{
\item{coco.object}{\code{(S4)} a fitted coco S4 object.}

\item{inv.hess}{\code{(matrix)} Inverse of the Hessian.}
}
\value{
the modified inverse of the hessian matrix
}
\description{
Compute confidence intervals for a (fitted) coco object
}
\author{
Federico Blasi
}
