% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getModHess}
\alias{getModHess}
\title{Retrieves the modified inverse of the hessian}
\usage{
getModHess(coco.object, inv.hess)
}
\arguments{
\item{coco.object}{\code{(S4)} a fitted coco S4 object.}

\item{inv.hess}{\code{(matrix)} Inverse of the Hessian.}
}
\value{
(\code{numeric matrix}) the modified inverse of the hessian matrix
}
\description{
Based on the inverse of the Hessian (based on the difference parameterization for the std.dev and scale parameters),
retrieves the modified inverse of the hessian (i.e. std.dev and scale).
}
\author{
Federico Blasi
}
