% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg2loglikelihood.R
\name{GetNeg2loglikelihood}
\alias{GetNeg2loglikelihood}
\title{GetNeg2loglikelihood}
\usage{
GetNeg2loglikelihood(theta, par.pos, locs, x_covariates, 
smooth.limits, z, n, lambda)
}
\arguments{
\item{theta}{a vector with parameters values}

\item{par.pos}{par.pos list}

\item{locs}{spatial location matrix}

\item{x_covariates}{design matrix}

\item{smooth.limits}{smooth.limits}

\item{z}{a vector of observed values}

\item{n}{dim(z)[1]}

\item{lambda}{regularization parameter}
}
\value{
value
}
\description{
compute the negative 2 log likelihood based on theta
}
\author{
Federico Blasi
}
