% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg2loglikelihood.R
\name{GetNeg2loglikelihoodTaper}
\alias{GetNeg2loglikelihoodTaper}
\title{GetNeg2loglikelihoodTaper}
\usage{
GetNeg2loglikelihoodTaper(theta, par.pos, ref_taper, locs, 
x_covariates, smooth.limits, cholS, z, n, lambda)
}
\arguments{
\item{theta}{a vector with parameters values}

\item{par.pos}{par.pos list}

\item{ref_taper}{spam object based on a taper based covariance function}

\item{locs}{spatial location matrix}

\item{x_covariates}{design matrix}

\item{smooth.limits}{smooth.limits}

\item{cholS}{Cholesky object from spam}

\item{z}{a vector of observed values}

\item{n}{dim(z)[1]}

\item{lambda}{regularization parameter}
}
\value{
value
}
\description{
compute the negative 2 log likelihood based on theta
}
\author{
Federico Blasi
}
