% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg2loglikelihood.R
\name{GetNeg2loglikelihoodTaper}
\alias{GetNeg2loglikelihoodTaper}
\title{GetNeg2loglikelihoodTaper}
\usage{
GetNeg2loglikelihoodTaper(theta, par.pos, ref_taper, locs, 
x_covariates, smooth.limits, cholS, z, n, lambda)
}
\arguments{
\item{theta}{\code{(numeric vector)} a vector with parameters values.}

\item{par.pos}{\code{(list)} par.pos list from \link{getDesignMatrix}.}

\item{ref_taper}{\code{(S4)} spam object based on a compact-supported covariance function.}

\item{locs}{\code{(matrix)} spatial location matrix.}

\item{x_covariates}{\code{(data.frame)} design matrix.}

\item{smooth.limits}{\code{(numeric vector)} smooth.limits.}

\item{cholS}{\code{(S4)} Cholesky object from spam.}

\item{z}{\code{(numeric vector)} a vector of observed values.}

\item{n}{\code{(numeric)} dim(z)[1].}

\item{lambda}{\code{(numeric)} regularization parameter.}
}
\value{
value
}
\description{
compute the negative 2 log likelihood based on theta
}
\author{
Federico Blasi
}
