% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getSpatEffects}
\alias{getSpatEffects}
\title{Computes the spatially-varying functions from a coco object}
\usage{
getSpatEffects(coco.object)
}
\arguments{
\item{coco.object}{\code{(S4)} a fitted coco S4 object.}
}
\value{
(\code{list}) a list with the different estimated surfaces.
}
\description{
Evaluates the spatially-varying functions of the nonstationary spatial structure.
}
\author{
Federico Blasi
}
