% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getSpatEffects}
\alias{getSpatEffects}
\title{Retrieves the estimated spatial effects of the spatial structure}
\usage{
getSpatEffects(coco.object)
}
\arguments{
\item{coco.object}{an coco class (fitted) object}
}
\value{
a list with the different estimated surfaces
}
\description{
Retrieves the estimated surfaces for different sources of nonstationarity
}
\author{
Federico Blasi
}
