% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getTrend}
\alias{getTrend}
\title{Computes the trend of the svcov object}
\usage{
getTrend(svcov.object)
}
\arguments{
\item{svov.object}{an svcov class (fitted) object}
}
\value{
a vector with the adjusted trend
}
\description{
Compute the trend of the (fitted) svcov object
}
\author{
Federico Blasi
}
