% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctions.R
\name{getTrend}
\alias{getTrend}
\title{Computes the trend of the coco object}
\usage{
getTrend(coco.object)
}
\arguments{
\item{coco.object}{a coco class (fitted) object}
}
\value{
a vector with the adjusted trend
}
\description{
Compute the trend of the (fitted) coco object
}
\author{
Federico Blasi
}
