% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holes}
\alias{holes}
\title{Holes Data Set}
\format{
A list with training and test data.frame with rows and variables:
\describe{
\item{x}{first spatial coordinate}
\item{y}{second spatial coordinate}
\item{cox_x}{first spatial characteristic}
\item{cov_y}{second spatial characteristic}
\item{z}{response variable}
}
}
\source{
Source of the data
}
\usage{
holes
}
\description{
The synthetic "holes" provides a set of training and test data.frame of a Gaussian process realization with a (inherently dense) nonstationary covariance function.
Four holes are present in the training dataset, and the task is to predict them.
}
\examples{
data(holes)
}
\keyword{datasets}
