% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holes_bm}
\alias{holes_bm}
\title{Holes with trend + multiple realizations Data Set}
\format{
A list with training, training.z, test, and test.z data.frames with rows and variables:
\describe{
\item{x}{first spatial coordinate}
\item{y}{second spatial coordinate}
\item{cox_x}{first spatial characteristic}
\item{cov_y}{second spatial characteristic}
\item{cov_z}{third spatial characteristic}
\item{z.i}{i-th response variable}
}
}
\source{
Source of the data
}
\usage{
holes_bm
}
\description{
The synthetic "holes_bm" provides a set of training and test data.frame of a Gaussian process realization with a (inherently dense) nonstationary covariance function.
Four holes are present in the training dataset, and the task is to predict them. This version provides ten independent realizations of the process, as well as considers
a spatial mean effect.
}
\examples{
data(holes_bm)
}
\keyword{datasets}
