% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFunctions.R
\name{is.formula}
\alias{is.formula}
\title{check whether an object belongs to a formula class}
\usage{
is.formula(x)
}
\arguments{
\item{x}{an R object}
}
\value{
TRUE/FALSE
}
\description{
check whether an object belongs to a formula class
}
\author{
Federico Blasi
}
