% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot,coco,missing-method}
\alias{plot,coco,missing-method}
\alias{plot,coco-method}
\title{Plot Method for Coco Class}
\usage{
\S4method{plot}{coco,missing}(x, y, ..., type = NULL, index = NULL, factr = 0.1, plot.control = NULL)
}
\arguments{
\item{x}{An object of class \code{coco}.}

\item{y}{Not used.}

\item{...}{Additional arguments passed to the plot function. when type "ellipse" , delta of nearest.dist must be specified.}

\item{type}{The type of plot. NULL or "ellipse" for drawing ellipse of the convolution kernels.}

\item{index}{For plotting local correlation plots.}

\item{factr}{Factor rate for size of ellipses.}

\item{plot.control}{Additional plot control parameters.}
}
\value{
A plot is created.
}
\description{
This method plots objects of class \code{coco}.
}
