% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,coco-method}
\title{Summary Method for Coco Class}
\usage{
\S4method{summary}{coco}(object, inv.hess = NULL)
}
\arguments{
\item{object}{(\code{S4}) An object of class 'coco'.}

\item{inv.hess}{(\code{numeric matrix} or \code{NULL}) inverse of the approximated hessian matrix (getHessian)}
}
\value{
summary the coco object
}
\description{
method summary for objects of class 'coco'.
}
\author{
Federico Blasi
}
